squash a -Wfloat-conversion due to sqrt in distBetweenPts
There is a dedicated libc function for doing square root of floats, but this
code was incorrectly calling the version for doubles instead. This updated code
is more appropriate and could even be more efficient.

+++ lib/glcomp/glutils.c
@@ -721,7 +721,7 @@ GLfloat distBetweenPts(glCompPoint A,glCompPoint B,float R)
 {
     GLfloat rv=0;	
     rv=(A.x-B.x)*(A.x-B.x) + (A.y-B.y)*(A.y-B.y) +(A.z-B.z)*(A.z-B.z);
-    rv=sqrt(rv);
+    rv=sqrtf(rv);
     if (rv <=R)
 	return 0;
     return rv;
